Parser declaraciones listo
