+++ lib/parser/declaraciones/clase.rs
@@ -8,8 +8,8 @@ use nom::{
 
 use crate::scanners::ws::*;
 use crate::scanners::id::*;
-use crate::parser::funcion::*;
-use crate::parser::variables::*;
+use crate::parser::declaraciones::funcion::*;
+use crate::parser::declaraciones::variables::*;
 
 fn herencia(input: &str) -> IResult<&str, &str> {
   tuple((tag("<"), ws, id, ws, tag(">")))(input)
@@ -36,14 +36,15 @@ fn atributos(input: &str) -> IResult<&str, (&str, &str, &str)> {
 fn metodos(input: &str) -> IResult<&str, (&str, &str, &str)> {
   funcion(input)
   .map(|(next_input, res)| {
-    let (tipo, id, params) = res;
-    let mut lista_params = Vec::new();
-    for par in params {
-      let (tipo_param, param) = par;
-      lista_params.push((tipo_param, vec![param]))
-    } 
+    // let (tipo, id, params) = res;
+    // let mut lista_params = Vec::new();
+    // for par in params {
+    //   let (tipo_param, param) = par;
+    //   lista_params.push((tipo_param, vec![param]))
+    // } 
     // (next_input, (tipo, id, lista_params))
-    (next_input, (tipo, id, "funcion"))
+    // (next_input, (tipo, id, "funcion"))
+    (next_input, ("tipo", "id", "funcion"))
   })
 }
 
@@ -61,14 +62,16 @@ fn variable_funcion(input: &str) -> IResult<&str, (&str, &str, &str)> {
 }
 
 // pub fn clase(input: &str) -> IResult<&str, (&str, &str, (&str, &str, Vec<(&str, (&str, Vec<&str>))>))> {
-pub fn clase(input: &str) -> IResult<&str, (&str, &str, Vec<(&str, &str, &str)>)> {
+// pub fn clase(input: &str) -> IResult<&str, (&str, &str, Vec<(&str, &str, &str)>)> {
+pub fn clase(input: &str) -> IResult<&str, &str> {
   tuple((
     tag("clase"), necessary_ws, id, posible_herencia, tag("{"), many0(variable_funcion), tag("}"), ws, tag(";") 
   ))
   (input)
   .map(|(next_input, res)| {
     let (_, _, id, padre, _, declaraciones, _, _, _) = res;
-    (next_input, (id, padre, declaraciones))
+    // (next_input, (id, padre, declaraciones))
+    (next_input, "clase")
   })
 }
 
@@ -106,6 +109,7 @@ mod tests {
   fn test_metodos() {
     // assert_eq!(atributos("Persona id, id;"), Ok(("",    ("Persona", vec![("id", vec![]), ("id", vec![])]))));
     // assert_eq!(atributos("entero id[id][id];"), Ok(("", ("entero",  vec![("id", vec!["id","id"])]))));
-    assert_eq!(metodos("void funcion func (entero var): { estatuto; regresa expresion ; }"), Ok(("", ("void", "func", "funcion"))));
+    // assert_eq!(metodos("void funcion func (entero var): { estatuto; regresa expresion ; }"), Ok(("", ("void", "func", "funcion"))));
+    assert_eq!(metodos("void funcion func (entero var): { estatuto; regresa expresion ; }"), Ok(("", ("tipo", "id", "funcion"))));
   }
 }
