+++ lib/parser/declaraciones/mod.rs
@@ -1,0 +1,4 @@
+pub mod clase;
+pub mod declaraciones;
+pub mod funcion;
+pub mod variables;
