+++ lib/parser/mod.rs
@@ -1,18 +1,15 @@
 // Folders
 pub mod reglas_expresion;
+pub mod declaraciones;
 
 // Archivos sueltos
 pub mod asignacion;
 pub mod bloque;
-pub mod clase;
 pub mod comentario;
 pub mod decision;
 pub mod dimensiones;
-pub mod declaraciones;
-pub mod funcion;
 pub mod func_esp;
 pub mod func_params;
 pub mod llama_func;
 pub mod programa;
 pub mod repeticion;
-pub mod variables;
